Rewriting code to consolidate into a single plugin. This reduces the need to construct intermediate cubes. Still missing are CLI tests associated with this new functionality.
