+++ lib/improver/nowcasting/accumulation.py
@@ -30,8 +30,8 @@
 # POSSIBILITY OF SUCH DAMAGE.
 """
 This module defines the Accumulation class for calculating precipitation
-accumulations from advected radar fields. It also includes the
-AccumulationAggregator class for combining these into longer periods.
+accumulations from advected radar fields. It is also possible to create longer
+accumulations from shorter intervals.
 """
 import warnings
 import numpy as np
@@ -45,12 +45,12 @@ from improver.utilities.cube_units import (enforce_coordinate_units_and_dtypes,
 class Accumulation:
     """
     Class to calculate precipitation accumulations from radar rates fields
-    provided at discrete time intervals. The plugin will return one fewer
-    cubes of accumulation than the number of input precipitation rate cubes.
-    The time intervals between the precipitation rate cubes determine the
-    accumulation periods.
+    provided at discrete time intervals. The plugin will calculate
+    accumulations between each pair of rates fields provided. These will be
+    used to construct the accumulation period requested when possible, and
+    cubes of this desired period are then returned.
     """
-    def __init__(self, accumulation_units='m'):
+    def __init__(self, accumulation_units='m', accumulation_period=None):
         """
         Initialise the plugin.
 
@@ -58,13 +58,21 @@ class Accumulation:
             accumulation_units (str):
                 The physical units in which the accumulation should be
                 returned. The default is metres.
+            accumulation_period (int):
+                The desired accumulation period in seconds. This period
+                must be evenly divisible by the time intervals of the input
+                cubes. The default is None, in which case an accumulation is
+                calculated across the span of time covered by the input rates
+                cubes.
         """
         self.accumulation_units = accumulation_units
+        self.accumulation_period = accumulation_period
 
     def __repr__(self):
         """Represent the plugin instance as a string."""
-        result = '<Accumulation: accumulation_units={}>'
-        return result.format(self.accumulation_units)
+        result = ('<Accumulation: accumulation_units={}, '
+                  'accumulation_period={}>')
+        return result.format(self.accumulation_units, self.accumulation_period)
 
     @staticmethod
     def sort_cubes_by_time(cubes):
@@ -87,6 +95,60 @@ class Accumulation:
         cubes = iris.cube.CubeList(np.array(cubes)[time_sorted])
         return cubes, times
 
+    def get_period_sets(self, time_interval, cubes):
+        """
+        Return sub-sets of the input cube list that will be used to construct
+        the accumulation period specified by self.accumulation_period.
+
+        Args:
+            time_interval (int):
+                The time interval between the input precipitation rate cubes in
+                seconds.
+            cubes (list):
+                The input precitation rate cubes.
+        Returns:
+            cube_subsets (list of lists):
+                A list containing lists which comprise the rates cubes required
+                to construct a sequence of accumulation cubes that match the
+                period given by self.accumulation_period. If
+                self.accumulation_period is None then the whole list of cubes
+                is returned to give one accumulation across the whole set.
+        Raises:
+            ValueError: Accumulation period cannot be created from time
+                        intervals of input cubes.
+        """
+        # If no accumulation period is provided, calculate accumulation across
+        # all provided cubes.
+        if self.accumulation_period is None:
+            return [cubes]
+
+        fraction, cube_subset = np.modf(self.accumulation_period /
+                                        time_interval)
+        if fraction != 0:
+            msg = ("The specified accumulation period ({}) is not divisible "
+                   "by the time intervals between rates cubes ({}). As "
+                   "a result it is not possible to calculate the desired "
+                   "total accumulation period.".format(
+                       self.accumulation_period, time_interval))
+            raise ValueError(msg)
+
+        cube_subset = int(cube_subset)
+
+        # We expect 1 more rates cubes than accumulations cubes, if we have
+        # more or less, a targetted accumulation_period will be incomplete.
+        unused_cubes = int(len(cubes) % cube_subset)
+        if unused_cubes != 1 and cube_subset != 1:
+            msg = ("The provided cubes result in a partial period given the "
+                   "specified accumulation_period, i.e. the number of cubes "
+                   "is insufficient to give a set of complete periods. Only "
+                   "complete periods will be returned.")
+            warnings.warn(msg)
+
+        cube_subsets = []
+        for index in range(0, len(cubes) - unused_cubes, cube_subset):
+            cube_subsets.append(cubes[index:index + cube_subset + 1])
+        return cube_subsets
+
     def process(self, cubes):
         """
         Calculate period precipitation accumulations based upon precipitation
@@ -98,8 +160,8 @@ class Accumulation:
         Returns:
             accumulation_cubes (iris.cube.CubeList):
                 A cubelist containing precipitation accumulation cubes where
-                the accumulation periods are determined by the intervals
-                between the precipitation rate cubes.
+                the accumulation periods are determined by plugin argument
+                accumulation_period.
         """
         # Standardise inputs to expected units.
         cubes = enforce_coordinate_units_and_dtypes(
@@ -109,167 +171,38 @@ class Accumulation:
 
         # Sort cubes into time order and calculate intervals.
         cubes, times = self.sort_cubes_by_time(cubes)
-        time_intervals = np.diff(times, axis=0)
-        accumulation_cubes = iris.cube.CubeList()
-
-        # Assumes rates cubes bookend the accumulations, e.g. there is one
-        # more rates cube than expected accumulation cubes.
-        iterator = zip(cubes[:-1], cubes[1:], time_intervals)
-
-        # Calculate accumulations and convert to desired units.
-        for start_cube, end_cube, time_interval in iterator:
-
-            # Use CubeCombiner to expand coordinate bounds.
-            cube_name = 'lwe_thickness_of_precipitation_amount'
-            accumulation_cube = CubeCombiner('add').process(
-                iris.cube.CubeList([start_cube, end_cube]), cube_name,
-                expanded_coord={'time': 'upper', 'forecast_period': 'upper'})
-            accumulation_cube.units = 'm'
-
-            # Calculate new data and insert into cube.
-            accumulation_data = start_cube.data * time_interval
-            accumulation_cube.data = accumulation_data
-            accumulation_cube.convert_units(self.accumulation_units)
-            accumulation_cubes.append(accumulation_cube)
-
-        return accumulation_cubes
-
-
-class AccumulationAggregator:
-
-    """Aggregate accumulations over short periods into longer periods."""
-
-    def __init__(self, accumulation_period=None):
-        """
-        Initialise the plugin.
-
-        Args:
-            accumulation_period (int):
-                The desired accumulation period in seconds. This period
-                must be evenly divisible by the time intervals of the input
-                cubes. The default is None, in which case an aggregation is
-                made across all input cubes.
-        """
-        self.accumulation_period = accumulation_period
-
-    def __repr__(self):
-        """Represent the plugin instance as a string."""
-        result = '<AccumulationAggregator: accumulation_period={}>'
-        return result.format(self.accumulation_period)
-
-    def check_accumulation_period(self, time_interval, ncubes):
-        """
-        Check to ensure the requested accumulation period can be created from
-        the input cubes.
-
-        Args:
-            time_interval (int):
-                The time interval between the input cubes in seconds.
-            ncubes (int):
-                The number of input cubes.
-        Returns:
-            cube_subset (int) or ncubes (int):
-                The integer number of cubes that must be combined to create the
-                desired total accumulation period, as given by
-                self.accumulation_period. Is self.accumulation_period is None
-                then ncubes is returned as the whole list will be aggregated.
-        Raises:
-            ValueError: Accumulation period cannot be created from time
-                        intervals of input cubes.
-        """
-        if self.accumulation_period is None:
-            return ncubes
-
-        fraction, cube_subset = np.modf(self.accumulation_period /
-                                        time_interval)
-        if fraction != 0:
-            msg = ("The specified accumulation period ({}) is not divisible "
-                   "by the time intervals between accumulation cubes ({}). As "
-                   "a result it is not possible to calculate the desired "
-                   "accumulation.".format(
-                       self.accumulation_period, time_interval))
-            raise ValueError(msg)
-
-        if ncubes % cube_subset != 0:
-            msg = ("The provided cubes result in a partial period given the "
-                   "specified accumulation_period, i.e. the number of cubes "
-                   "is insufficient to give a set of complete periods. Only "
-                   "complete periods will be returned.")
-            warnings.warn(msg)
-
-        return int(cube_subset)
-
-    @staticmethod
-    def chunk_list(cubes, cube_subset):
-        """
-        Break the CubeList into a set of CubeLists each containing the
-        required number of cubes to be combined to create the desired
-        accumulation_period.
-
-        Args:
-            cubes (iris.cube.CubeList):
-                The input cubes.
-            cube_subset (int):
-                The integer number of cubes that must be combined to create the
-                desired total accumulation period, as given by
-                self.accumulation_period.
-        Returns:
-            generator:
-                A generator that produces the short sub-lists required to
-                create new accumulation periods.
-        """
-        for index in range(0, len(cubes), cube_subset):
-            yield iris.cube.CubeList(cubes[index:index + cube_subset])
-
-    def process(self, cubes):
-        """
-        Calculate aggregated accumulations over the desired accumulation period
-        after checking this is possible given the input cubes.
-
-        Args:
-            cubes (iris.cube.CubeList):
-                A cubelist containing input precipitation accumulation cubes.
-        Returns:
-            accumulation_cubes (iris.cube.CubeList):
-                A cubelist containing precipitation accumulation cubes where
-                the accumulation periods are determined by the given
-                accumulation_period unless this is None, in which case a single
-                cube will be returned covering the entire period of the input
-                cubes.
-        """
-        # Sort cubes into time order and calculate intervals.
-        cubes, times = Accumulation.sort_cubes_by_time(cubes)
-        ncubes = len(cubes)
-
-        # Find the common time interval between accumulation cubes.
         try:
             time_interval, = np.unique(np.diff(times, axis=0))
         except ValueError:
-            msg = ("AccumulationAggregator is designed to work with "
-                   "accumulation cubes at regualar time intervals. Cubes "
+            msg = ("Accumulation is designed to work with "
+                   "rates cubes at regualar time intervals. Cubes "
                    "provided are unevenly spaced in time; time intervals are "
                    "{}.".format(np.diff(times, axis=0)))
             raise ValueError(msg)
 
-        cube_subset_no = self.check_accumulation_period(time_interval, ncubes)
-        cubes_to_aggregate = self.chunk_list(cubes, cube_subset_no)
+        cube_subsets = self.get_period_sets(time_interval, cubes)
 
-        aggregated_cubes = iris.cube.CubeList()
-        for cube_subset in cubes_to_aggregate:
-
-            # Don't produce an accumulation cube for any incomplete periods.
-            if len(cube_subset) != cube_subset_no:
-                break
-
-            # If a single cube period matches the expected period, do nothing
+        accumulation_cubes = iris.cube.CubeList()
+        for cube_subset in cube_subsets:
+            # Handle accumulation periods equal to the rates cube time interval
+            # in which case the last subset does not get used.
             if len(cube_subset) == 1:
-                aggregated_cubes.append(cube_subset[0])
                 continue
 
-            # Aggregate shorter period accumulations and expand coord bounds.
-            aggregated_cube = CubeCombiner('add').process(
-                cube_subset, cube_subset[0].name(),
-                expanded_coord={'time': 'upper', 'forecast_period': 'upper'})
-            aggregated_cubes.append(aggregated_cube)
+            accumulation = 0.
+            for cube in cube_subset[0:-1]:
+                accumulation += cube.data * time_interval
 
-        return aggregated_cubes
+            # Use CubeCombiner to expand coordinate bounds.
+            cube_name = 'lwe_thickness_of_precipitation_amount'
+            accumulation_cube = CubeCombiner('add').process(
+                iris.cube.CubeList(cube_subset), cube_name,
+                expanded_coord={'time': 'upper', 'forecast_period': 'upper'})
+            accumulation_cube.units = 'm'
+
+            # Calculate new data and insert into cube.
+            accumulation_cube.data = accumulation
+            accumulation_cube.convert_units(self.accumulation_units)
+            accumulation_cubes.append(accumulation_cube)
+
+        return accumulation_cubes
